refactor(experimental): graphql: group schema type resolvers
