+++ packages/rpc-graphql/src/schema/type-resolvers/block.ts
@@ -1,0 +1,8 @@
+import { BlockResult } from '../../resolvers/block';
+
+export const blockTypeResolvers = {
+    Block: {
+        transactions: (parent?: BlockResult) =>
+            parent?.transactionResults ? Object.values(parent.transactionResults) : null,
+    },
+};
