+++ packages/rpc-graphql/src/schema/type-resolvers/transaction.ts
@@ -1,0 +1,7 @@
+import { resolveTransactionData } from '../../resolvers/transaction';
+
+export const transactionTypeResolvers = {
+    Transaction: {
+        data: resolveTransactionData(),
+    },
+};
